🚚 Add "Model" suffix to all mutation type models
