+++ app/src/main/java/com/suvanl/fixmylinks/domain/mutation/model/AllUrlParamsMutationModel.kt
@@ -2,15 +2,10 @@ package com.suvanl.fixmylinks.domain.mutation.model
 
 import com.suvanl.fixmylinks.domain.mutation.MutationType
 
-data class SpecificUrlParamsMutationInfo(
-    val removableParams: List<String>
-)
-
-data class SpecificUrlParamsMutation(
+data class AllUrlParamsMutationModel(
     override val name: String,
     override val mutationType: MutationType,
     override val triggerDomain: String,
     override val dateModifiedTimestamp: Long? = null,
-    override val isLocalOnly: Boolean,
-    val mutationInfo: SpecificUrlParamsMutationInfo,
-) : BaseMutation
+    override val isLocalOnly: Boolean
+) : BaseMutationModel
