+++ app/src/main/java/com/suvanl/fixmylinks/domain/mutation/model/DomainNameAndAllUrlParamsMutationModel.kt
@@ -2,17 +2,11 @@ package com.suvanl.fixmylinks.domain.mutation.model
 
 import com.suvanl.fixmylinks.domain.mutation.MutationType
 
-data class DomainNameAndSpecificUrlParamsMutationInfo(
-    val initialDomainName: String,
-    val targetDomainName: String,
-    val removableParams: List<String>,
-)
-
-data class DomainNameAndSpecificUrlParamsMutation(
+data class DomainNameAndAllUrlParamsMutationModel(
     override val name: String,
     override val mutationType: MutationType,
     override val triggerDomain: String,
     override val dateModifiedTimestamp: Long? = null,
     override val isLocalOnly: Boolean,
-    val mutationInfo: DomainNameAndSpecificUrlParamsMutationInfo,
-) : BaseMutation
+    val mutationInfo: DomainNameMutationInfo
+) : BaseMutationModel
