+++ app/src/main/java/com/suvanl/fixmylinks/domain/mutation/model/DomainNameMutationModel.kt
@@ -2,11 +2,16 @@ package com.suvanl.fixmylinks.domain.mutation.model
 
 import com.suvanl.fixmylinks.domain.mutation.MutationType
 
-data class DomainNameAndAllUrlParamsMutation(
+data class DomainNameMutationInfo(
+    val from: String,
+    val to: String
+)
+
+data class DomainNameMutationModel(
     override val name: String,
-    override val mutationType: MutationType,
+    override val mutationType: MutationType = MutationType.DOMAIN_NAME,
     override val triggerDomain: String,
     override val dateModifiedTimestamp: Long? = null,
     override val isLocalOnly: Boolean,
-    val mutationInfo: DomainNameMutationInfo
-) : BaseMutation
+    val mutationInfo: DomainNameMutationInfo,
+) : BaseMutationModel
